pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "map-pfn"
version = "0.0.1"
requires-python = "==3.12.*"
dependencies = [
  "jax[cuda12]==0.5.3",
  "torch==2.8.0",
  "wandb>=0.18.5",
  "hydra-zen>=0.14.0",
  "submitit>=1.5.1",
  "python-dotenv>=1.0.1",
  "numpy>=1.26.4",
  "pydantic>=2.10.3",
  "pytest>=8.3.5",
  "pandas>=2.3.1",
  "scikit-learn>=1.7.1",
  "scipy>=1.16.1",
  "pyarrow>=21.0.0",
  "pytorch-lightning>=2.5.1",
  "matplotlib>=3.8.1",
  "anndata==0.10.8",
  "equinox>=0.13.0",
  "optax>=0.2.5",
  "diffrax>=0.7.0",
  "jaxtyping>=0.3.2",
  "networkx>=3.5",
  "seaborn>=0.13.2",
  "ott-jax>=0.5.1",
  "ipykernel>=6.30.1",
  "einops>=0.8.1",
  "huggingface-hub>=0.34.3",
  "igraph>=0.11.9",
  "datasets>=4.4.1",
  "perturbench>=0.0.1",
  "ml-project-template>=0.0.14",
  "click>=8.1.7",
  "pertpy>=1.0.3",
  "mudata==0.3.0",
  "sergio-rs>=0.2.1",
]

[dependency-groups]
baselines = [
  "hydra-colorlog==1.2.0",
  "hydra-core==1.3.2",
  "hydra-joblib-launcher==1.2.0",
  "hydra-submitit-launcher==1.2.0",
  "matplotlib==3.8.1",
  "omegaconf==2.3.0",
  "pillow==10.1.0",
  "pot==0.9.4",
  "python-dotenv==1.0.1",
  "pyyaml==6.0.2",
  "rich==13.7.1",
  "torch-cluster==1.6.3",
  "torch-geometric==2.5.3",
  "torchdyn==1.0.6",
]
dev = ["notebook>=7.4.7", "pre-commit>=4.0.1", "ruff>=0.8.0"]

[tool.setuptools.packages.find]
include = ["map-pfn", "baselines"]
namespaces = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torch-cluster = { url = "https://data.pyg.org/whl/torch-2.8.0%2Bcu128/torch_cluster-1.6.3%2Bpt28cu128-cp312-cp312-linux_x86_64.whl" }
ml-project-template = { git = "https://github.com/marvinsxtr/ml-project-template.git", branch = "main" }
diffrax = { git = "https://github.com/patrick-kidger/diffrax.git", branch = "main" }
perturbench = { git = "https://github.com/marvinsxtr/perturbench.git", branch = "main" }
sergio_rs = { url = "https://github.com/marvinsxtr/sergio_rs/releases/download/v0.2.1/sergio_rs-0.2.1-cp312-cp312-linux_x86_64.whl" }

[[tool.uv.index]]
explicit = true
name = 'pytorch-cu128'
url = 'https://download.pytorch.org/whl/cu128'

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
explicit_package_bases = true
disable_error_code = ["import-untyped"]

[tool.ruff]
line-length = 119
indent-width = 4
extend-exclude = [
  "**/notebooks",
  "**/tests",
  "**/baselines",
  "map_pfn/data/grn",
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN002",
  "ANN003",
  "ANN401",
  "D413",
  "COM812",
  "D100",
  "D104",
  "D107",
  "D205",
  "PD901",
  "D400",
  "D401",
  "D415",
  "FA",
  "SLF",
  "INP",
  "TRY003",
  "TRY201",
  "EM",
  "FBT",
  "RET",
  "C406",
  "E741",
  "PLR2004",
  "RUF009",
  "RUF012",
  "BLE001",
  "S603",
  "S607",
  "S506",
  "FIX002",
  "NPY002",
  "G004",
  "S311",
  "PIE790",
  "TRY400",
  "S108",
  "W191",
  "E111",
  "E114",
  "E117",
  "D206",
  "D300",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "COM812",
  "COM819",
  "D203",
  "D213",
  "N806",
  "N803",
  "E712",
  "PLR0913",
  "TC001",
  "I001",
  "F722",
  "TC003",
  "PD008",
  "TC002",
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = ["S101", "ARG", "FBT"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"pytorch_lightning".msg = "Use lightning.pytorch instead to avoid type validation issues."
